Rename manuscripts2/new_functions.py -> manuscripts2/elasticsearch.py Move PullRequests and Issues classes to elasticsearch.py Delete mansucripts2/derived_classes.py
